Вынести Регистрацию студентов в отдельный файл
