+++ vk_bot/functions/notifications.py
@@ -32,3 +32,14 @@ def calculating_reminder_times(schedule, time: int) -> list:
     }
 
     return reminders
+
+
+def get_notifications_status(time):
+    """Статус напоминаний"""
+    if not time or time == 0:
+        notifications_status = 'Напоминания выключены ❌\n' \
+                               'Воспользуйтесь настройками, чтобы включить'
+    else:
+        notifications_status = f'Напоминания включены ✅\n' \
+                               f'Сообщение придёт за {time} мин до начала пары 😇'
+    return notifications_status
